fix: Extract features into folder
