+++ .storybook/preview.js
@@ -6,11 +6,11 @@ import { useTranslation } from 'react-i18next';
 import { useDarkMode } from 'storybook-dark-mode';
 
 import { Providers } from '../src/app/Providers';
-import i18nGlobal from '../src/config/i18next';
+import i18nGlobal from '../src/config/i18n/i18next';
 import {
   AVAILABLE_LANGUAGES,
   DEFAULT_LANGUAGE_KEY,
-} from '../src/constants/i18n';
+} from '../src/lib/i18n/constants';
 import logoReversed from './logo-reversed.svg';
 import logo from './logo.svg';
