--- src/app/(app)/account/account.types.ts
@@ -1,3 +1,0 @@
-import { User } from '@/app/(app)/admin/users/users.types';
-
-export type Account = User;
