+++ src/app/(app)/account/password/page.tsx
@@ -7,11 +7,11 @@ import { Formiz, useForm } from '@formiz/core';
 import { isMaxLength, isMinLength } from '@formiz/validations';
 import { useTranslation } from 'react-i18next';
 
-import { AccountNav } from '@/app/(app)/account/AccountNav';
-import { useUpdatePassword } from '@/app/(app)/account/account.service';
 import { FieldInput } from '@/components/FieldInput';
 import { Page, PageContent } from '@/components/Page';
 import { useToastError, useToastSuccess } from '@/components/Toast';
+import { AccountNav } from '@/features/account/AccountNav';
+import { useUpdatePassword } from '@/features/account/service';
 
 export default function PagePassword() {
   const { t } = useTranslation(['account']);
