+++ src/app/(app)/admin/api/page.tsx
@@ -6,8 +6,8 @@ import { Box } from '@chakra-ui/react';
 import SwaggerUI from 'swagger-ui-react';
 import 'swagger-ui-react/swagger-ui.css';
 
-import { AdminNav } from '@/app/(app)/admin/AdminNav';
 import { Page, PageContent } from '@/components/Page';
+import { AdminNav } from '@/features/admin/AdminNav';
 
 export default function PageApiDocumentation() {
   return (
