+++ src/app/(app)/logout/page.tsx
@@ -6,7 +6,7 @@ import { Center, Spinner } from '@chakra-ui/react';
 import { useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 
-import { useAuthContext } from '@/app/AuthContext';
+import { useAuthContext } from '@/features/auth/AuthContext';
 
 export default function PageLogout() {
   const { updateToken } = useAuthContext();
