+++ src/app/Providers.tsx
@@ -5,9 +5,11 @@ import { ChakraProvider } from '@chakra-ui/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 
-import { AuthProvider } from '@/app/AuthContext';
-import '@/config';
-import { AVAILABLE_LANGUAGES } from '@/constants/i18n';
+import { AuthProvider } from '@/features/auth/AuthContext';
+import '@/lib/axios/config';
+import '@/lib/dayjs/config';
+import '@/lib/i18n/config';
+import { AVAILABLE_LANGUAGES } from '@/lib/i18n/constants';
 import theme from '@/theme';
 
 const queryClient = new QueryClient();
