--- src/config/index.ts
@@ -1,3 +1,0 @@
-import './axios';
-import './dayjs';
-import './i18next';
