+++ src/features/account/types.ts
@@ -1,0 +1,3 @@
+import { User } from '@/features/users/types';
+
+export type Account = User;
